requirements.txt
numpy==1.17.0
Pillow==6.1.0
tqdm==4.35.0
